+++ digiplan/map/choropleths.py
@@ -206,7 +206,7 @@ class ElectricityDemandChoropleth(Choropleth):  # noqa: D101
 class ElectricityDemand2045Choropleth(Choropleth):  # noqa: D101
     def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
         return (
-            calculations.electricity_demand_per_municipality_2045(self.map_state["simulation_id"]).sum(axis=1).to_dict()
+            calculations.electricity_demand_per_municipality_2045(self.map_state["pre_result_id"]).sum(axis=1).to_dict()
         )
 
 
@@ -223,7 +223,7 @@ class ElectricityDemandCapita2045Choropleth(Choropleth):  # noqa: D101
     def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
         capita_demand = (
             calculations.calculate_capita_for_value(
-                calculations.electricity_demand_per_municipality_2045(self.map_state["simulation_id"]),
+                calculations.electricity_demand_per_municipality_2045(self.map_state["pre_result_id"]),
             ).sum(axis=1)
             * 1e6
         )
@@ -253,7 +253,7 @@ class HeatDemandCapita2045Choropleth(Choropleth):  # noqa: D101
     def get_values_per_feature(self) -> dict[int, float]:  # noqa: D102
         capita_demand = (
             calculations.calculate_capita_for_value(
-                calculations.heat_demand_per_municipality_2045(self.map_state["simulation_id"]),
+                calculations.heat_demand_per_municipality_2045(self.map_state["pre_result_id"]),
             ).sum(axis=1)
             * 1e6
         )